lint: add clang-tidy + cppcheck routine, expand warning set
- tools/lint.sh: configures build-lint/ with homebrew clang on first run,
  runs clang-tidy (parallel via run-clang-tidy) + cppcheck with curated
  suppressions for known false positives.
- .clang-tidy: bugprone/performance/portability/clang-analyzer/cert/misc
  minus low-signal noise (cert-err33-c stdio returns, narrowing-conversions,
  pragma-once portability, etc.).
- cmake target `lint` for cmake --build build --target lint (Mac only).
- CompilerWarnings.cmake: gate GCC-only -W flags behind GNU id so
  clang-tidy doesn't choke; add -Wuseless-cast (GCC) and the cpp-best-
  practices MSVC /w14xxx promotion set.
- Real findings fixed:
  - snes_io.cpp/genesis.cpp: argument-eval-order around Image ctor with
    size() + std::move; hoisted size() to a local.
  - ham.cpp: run_passes by value, hot push_back without reserve, stale
    std::move on by-ref param.
  - strips.cpp/main.cpp: 5 dead stores, 12 stale /*name=*/ comments.
  - ssimulacra2.cpp: mu1/mul confusable -> renamed mul->prod.
  - api.cpp: pl/p1 confusable identifiers (planes/palette-lab) renamed.
  - c64.cpp: too-small loop variable + static_assert.
  - 11 useless-cast sites cleaned up.
- All 178 ctests pass, output byte-identical.

+++ src/cheader.cpp
@@ -1075,7 +1075,7 @@ Result<std::string> generate_viewer(const bitplane::BitplaneData& planes,
         // AGA >32: double writes (LOCT high + low) + BPLCON3 per bank + reset
         auto num_banks = (pal_count + 31) / 32;
         cop_size += pal_count * 4;  // double the color writes (LOCT=1 pass)
-        cop_size += static_cast<std::size_t>((num_banks * 2 + 1) * 4);  // BPLCON3 switches
+        cop_size += (num_banks * 2 + 1) * 4;  // BPLCON3 switches
     } else if (options.aga) {
         // AGA <=32: LOCT=1 pass + BPLCON3 switches (2 for LOCT on/off)
         cop_size += pal_count * 4 + 8;
